refactor: refactored project to be more general
Refactored project to be as general as possible to add more functionality unrelated to adding. Also changed the error check to state the coming property, with arbitrary amount of properties.
