+++ config.ini
@@ -1,14 +1,17 @@
-[SETTINGS]
-
-# Output into a CSV instead of adding through Homebox API.
-# Homebox URL and login till required.
-output_into_csv = False
-
-# Let LLM generate a description based on item name.
-# Description may be wrong, mostly there for flavour.
-generate_description = False
+[VOICE_RECOGNITION]
 
 # Attempts conversion from your file format to WAV
 # Requires FFMPEG to be installed
 # Setting does not matter if you are only using WAV or FLAC files
 attempt_conversion = True
+
+[ADDER]
+
+# Let LLM generate a description based on item name.
+# Description may be wrong, mostly there for flavour.
+generate_description = False
+
+# Output into a CSV instead of adding through Homebox API.
+# Homebox URL and login still required.
+output_into_csv = False
+
